+++ packages/inertia-vue3/src/app.js
@@ -1,4 +1,4 @@
-import useForm from './form'
+import useForm from './useForm'
 import link from './link'
 import remember from './remember'
 import { computed, h, markRaw, ref } from 'vue'
