+++ packages/inertia-vue3/src/index.js
@@ -1,4 +1,4 @@
 export { default as App, default as app, plugin, usePage } from './app'
 export { default as Link, default as link } from './link'
 export { useRemember } from './remember'
-export { default as useForm } from './form'
+export { default as useForm } from './useForm'
